#3575 CA-1: migrar provider-error-parser a lib/agent-launcher + shim estatico
- Copia el parser a `.pipeline/lib/agent-launcher/provider-error-parser.js`,
  ajustando el unico require relativo afectado (`./anthropic-1m-workaround`
  -> `../commander/anthropic-1m-workaround`) sin tocar la API publica ni
  ninguna otra dependencia.
- Reemplaza `.pipeline/lib/commander/provider-error-parser.js` por un re-export
  shim estatico (`module.exports = require('../agent-launcher/...')`) literal,
  sin templates ni variables en el path (cumple NEW-5 de #3435).
- Emision unica por proceso de `console.warn` con mensaje literal (sin contexto
  dinamico) para indicar deprecacion del path legacy.
- No-op funcional: `multi-provider.js:713` y los tests del Commander resuelven
  el modulo por el shim sin cambios.

Verificacion DoD:
  - `node --test .pipeline/lib/commander/__tests__/provider-error-parser.test.js`
    -> 44/44 pass via shim.
  - `node --test .pipeline/lib/commander/__tests__/anthropic-1m-workaround.test.js`
    -> 31/31 pass.
  - `node -e "require(...); require(...);" | grep -c deprecated` -> 1.
  - `grep -E "require\(.*\\${|require\(.*\+|require\(process\.env"` sobre el
    shim no matchea nada (sin requires dinamicos).
